+++ paragone/align_selected_and_tree.py
@@ -7,8 +7,8 @@ Input is fasta files that have undergone QC processes; this QC may have removed 
 This script:
 
     - Checks for any paralogs in internal outgroups (if the latter are specified), and selects a single
-    representative sequence for each taxon  - this will be the sequence with the highest distance from a sample of up
-    to 10 ingroup sequences.
+      representative sequence for each taxon  - this will be the sequence with the highest distance from a sample of up
+      to 10 ingroup sequences.
     - Aligns the sequences using MAFFT.
     - Generates trees from the alignments using either IQTREE or FastTreeMP.
 
@@ -28,10 +28,10 @@ from collections import defaultdict
 from Bio import SeqIO, AlignIO
 from Bio.Phylo.TreeConstruction import DistanceCalculator
 from Bio.Align import MultipleSeqAlignment
-from Bio.Align.Applications import MafftCommandline, ClustalOmegaCommandline, MuscleCommandline
 from concurrent.futures.process import ProcessPoolExecutor
 from multiprocessing import Manager
-from concurrent.futures import wait
+from concurrent.futures import wait, as_completed
+import traceback
 
 from paragone import utils
 from paragone.align_and_clean import run_trimal
@@ -41,8 +41,6 @@ def add_outgroup_seqs(qc_alignment_directory,
                       selected_alignment_directory,
                       list_of_internal_outgroups,
                       list_of_external_outgroups,
-                      # file_of_external_outgroups,
-                      # list_of_external_outgroups=None,
                       logger=None):
     """
     Check the number of genes that have an outgroup sequence in either the list_of_internal_outgroups (i.e.
@@ -52,18 +50,16 @@ def add_outgroup_seqs(qc_alignment_directory,
 
     Write a tab-separated outgroup file for the Y&S pruning scripts, of the form:
 
-    IN  Euchiton_limosus
-    IN  Euchiton_sphaericus
-    IN  Pterochaeta_paniculata
-    OUT sunf
-    etc...
+        IN  Euchiton_limosus
+        IN  Euchiton_sphaericus
+        IN  Pterochaeta_paniculata
+        OUT sunf
+        etc...
 
     :param str qc_alignment_directory:
     :param str selected_alignment_directory:
     :param list list_of_internal_outgroups:
     :param list list_of_external_outgroups:
-    # :param str file_of_external_outgroups:
-    # :param list_of_external_outgroups:
     :param logging.Logger logger: a logger object
     :return str output_folder: name of output folder containing fasta with outgroups added
     """
@@ -71,7 +67,6 @@ def add_outgroup_seqs(qc_alignment_directory,
     logger.debug(f'list_of_internal_outgroups: {list_of_internal_outgroups}')
     logger.debug(f'list_of_external_outgroups_to_select: {list_of_external_outgroups}')
 
-    # if not file_of_external_outgroups and not list_of_internal_outgroups:
     if len(list_of_internal_outgroups) == 0 and len(list_of_external_outgroups) == 0:
         logger.warning(f'{"[WARNING]:":10} No external or internal outgroups supplied!')
 
@@ -82,10 +77,12 @@ def add_outgroup_seqs(qc_alignment_directory,
     # list_of_internal_outgroups:
     internal_outgroup_dict = defaultdict(lambda: defaultdict(list))
     all_paralog_taxon_names = set()
+    internal_outgroup_found_dict = dict()
 
     for original_alignment in glob.glob(f'{qc_alignment_directory}/*.fasta'):
         gene_id = os.path.basename(original_alignment).split('.')[0]  # get prefix e.g. '4471'
         seqs = SeqIO.parse(original_alignment, 'fasta')
+        internal_outgroup_found_dict[gene_id] = False  # set default
 
         # Populate the internal_outgroup_dict (potentially more than one sequence per taxon):
         for seq in seqs:
@@ -93,6 +90,7 @@ def add_outgroup_seqs(qc_alignment_directory,
             all_paralog_taxon_names.add(seq_name_prefix)
             if list_of_internal_outgroups and seq_name_prefix in list_of_internal_outgroups:
                 internal_outgroup_dict[gene_id][seq_name_prefix].append(seq)
+                internal_outgroup_found_dict[gene_id] = True
 
         if list_of_internal_outgroups:
             alignment = AlignIO.read(original_alignment, 'fasta')
@@ -122,6 +120,36 @@ def add_outgroup_seqs(qc_alignment_directory,
                                                                         logger=logger)
             # Reassign internal_outgroup_dict[gene_id] dictionary entry to selected sequences:
             internal_outgroup_dict[gene_id] = internal_outgroup_dict_filtered[gene_id]
+
+    # Check that internal outgroups were actually found in the "qc_alignment_directory" provided:
+    genes_with_internal_outgroups_found = [gene_id for gene_id in internal_outgroup_found_dict.keys() if
+                                           internal_outgroup_found_dict[gene_id]]
+    genes_with_no_internal_outgroups_found = [gene_id for gene_id in internal_outgroup_found_dict.keys() if not
+                                              internal_outgroup_found_dict[gene_id]]
+
+    logger.debug(f'genes_with_internal_outgroups_found: {genes_with_internal_outgroups_found}')
+    logger.debug(f'genes_with_no_internal_outgroups_found: {genes_with_no_internal_outgroups_found}')
+
+    if len(list_of_internal_outgroups) != 0 and len(genes_with_internal_outgroups_found) == 0:
+        fill = textwrap.fill(f'{"[ERROR]:":10} No internal outgroup sequences were found for any gene within the '
+                             f'directory provided: "{qc_alignment_directory}". Please check you have provided the '
+                             f'correct directory!',
+                             width=90, subsequent_indent=' ' * 11,
+                             break_on_hyphens=False)
+        logger.error(f'{fill}')
+        sys.exit()
+
+    if len(genes_with_no_internal_outgroups_found) != 0:
+        fill = textwrap.fill(f'{"[INFO]:":10} No internal outgroup sequences were found for some genes in the '
+                             f'directory "{qc_alignment_directory}". This may be expected i.e. if some of your input '
+                             f'paralog fasta files do not have any internal outgroup sequences. The genes are:',
+                             width=90, subsequent_indent=' ' * 11,
+                             break_on_hyphens=False)
+
+        logger.info(f'{fill}')
+
+        for gene_id in genes_with_no_internal_outgroups_found:
+            logger.info(f'{" " * 11}{gene_id}')
 
     # Read in external outgroups file if present, and create a dictionary of gene_id:list_of_seq_names, either for all
     # seqs if no external outgroup taxa specified, or for specified taxa only:
@@ -235,10 +263,19 @@ def filter_internal_outgroups(internal_outgroup_dict,
                 distance_values = trimmed_dm[seq.name]
                 sorted_distance_values = sorted(distance_values, key=float)
                 closest_sequence_distance = sorted_distance_values[1]  # skip zero as it's self-vs-self
-                if closest_sequence_distance > seq_to_keep_distance:
+
+                # Make sure a sequence is selected:
+                if not seq_to_keep:
+                    seq_to_keep_distance = closest_sequence_distance
+                    seq_to_keep = seq
+                    continue
+
+                # If a sequence has been selected already, compare it to the current seq:
+                if closest_sequence_distance > seq_to_keep_distance:  # i.e. it's more diverged
                     logger.debug(f'{closest_sequence_distance} is larger than {seq_to_keep_distance}')
                     seq_to_keep_distance = closest_sequence_distance
                     seq_to_keep = seq
+
             logger.debug(f'Keeping sequence {seq_to_keep.name} with distance {seq_to_keep_distance}')
             internal_outgroup_dict_copy[gene_id][taxon_id] = [seq_to_keep]
 
@@ -265,7 +302,7 @@ def mafft_align_multiprocessing(fasta_to_align_folder,
     output_folder = f'11_pre_paralog_resolution_alignments'
     utils.createfolder(output_folder)
 
-    logger.info(f'{"[INFO]:":10} Generating alignments for fasta files using MAFFT...')
+    logger.info(f'{"[INFO]:":10} Generating alignments from fasta files using MAFFT...')
 
     # Filter out any input files with fewer than four sequences:
     target_genes = []
@@ -293,8 +330,16 @@ def mafft_align_multiprocessing(fasta_to_align_folder,
                                       logger=logger)
                           for fasta_file in target_genes]
 
-        for future in future_results:
-            future.add_done_callback(utils.done_callback)
+        for future in as_completed(future_results):
+
+            try:
+                check = future.result()
+
+            except Exception as error:
+                print(f'Error raised: {error}')
+                tb = traceback.format_exc()
+                print(f'traceback is:\n{tb}')
+
         wait(future_results, return_when="ALL_COMPLETED")
 
     alignment_list = [alignment for alignment in glob.glob(f'{output_folder}/*.aln.fasta') if
@@ -342,19 +387,31 @@ def mafft_align(fasta_file,
     except AssertionError:
 
         if algorithm == 'auto':
-            mafft_cline = (MafftCommandline(auto='true', thread=threads, input=fasta_file))
+            command = f'mafft --auto --thread {threads} {fasta_file} > {expected_alignment_file}'
         else:
-            mafft_cline = (MafftCommandline(algorithm, thread=threads, input=fasta_file))
+            command = f'{algorithm} --thread {threads} {fasta_file} > {expected_alignment_file}'
 
-        logger.debug(f'{"[INFO]:":10} Performing MAFFT alignment with command: {mafft_cline}')
+        logger.debug(f'{"[INFO]:":10} Performing MAFFT alignment with command: {command}')
 
-        stdout, stderr = mafft_cline()
+        try:
 
-        # logger.debug(f'stdout is: {stdout}')
-        # logger.debug(f'stderr is: {stderr}')
+            result = subprocess.run(command,
+                                    universal_newlines=True,
+                                    stdout=subprocess.PIPE,
+                                    stderr=subprocess.PIPE,
+                                    check=True,
+                                    shell=True)
 
-        with open(expected_alignment_file, 'w') as alignment_file:
-            alignment_file.write(stdout)
+            logger.debug(f'MAFFT check_returncode() is: {result.check_returncode()}')
+            logger.debug(f'MAFFT stdout is: {result.stdout}')
+            logger.debug(f'MAFFT stderr is: {result.stderr}')
+
+        except subprocess.CalledProcessError as exc:
+            logger.error(f'MAFFT FAILED. Output is: {exc}')
+            logger.error(f'MAFFT stdout is: {exc.stdout}')
+            logger.error(f'MAFFT stderr is: {exc.stderr}')
+
+            raise ValueError('There was an issue running MAFFT. Check input files!')
 
         with lock:
             counter.value += 1
@@ -401,8 +458,17 @@ def clustalo_align_multiprocessing(fasta_to_align_folder,
                                       threads=clustalo_threads,
                                       logger=logger)
                           for fasta_file in target_genes]
-        for future in future_results:
-            future.add_done_callback(utils.done_callback)
+
+        for future in as_completed(future_results):
+
+            try:
+                check = future.result()
+
+            except Exception as error:
+                print(f'Error raised: {error}')
+                tb = traceback.format_exc()
+                print(f'traceback is:\n{tb}')
+
         wait(future_results, return_when="ALL_COMPLETED")
 
     alignment_list = [alignment for alignment in glob.glob(f'{output_folder}/*.aln.fasta') if
@@ -446,14 +512,28 @@ def clustalo_align(fasta_file,
         return os.path.basename(expected_alignment_file)
 
     except AssertionError:
-        clustalomega_cline = ClustalOmegaCommandline(infile=fasta_file, outfile=expected_alignment_file,
-                                                     verbose=True, auto=True, threads=threads)
-        logger.info(f'{"[INFO]:":10} Performing Clustal alignment with command: {clustalomega_cline}')
+        command = f'clustalo --in {fasta_file} --out {expected_alignment_file} --threads {threads} --verbose'
+        logger.info(f'{"[INFO]:":10} Performing Clustal alignment with command: {command}')
 
-        stdout, stderr = clustalomega_cline()
+        try:
 
-        # logger.debug(f'stdout is: {stdout}')
-        # logger.debug(f'stderr is: {stderr}')
+            result = subprocess.run(command,
+                                    universal_newlines=True,
+                                    stdout=subprocess.PIPE,
+                                    stderr=subprocess.PIPE,
+                                    check=True,
+                                    shell=True)
+
+            logger.debug(f'ClustalO check_returncode() is: {result.check_returncode()}')
+            logger.debug(f'ClustalO stdout is: {result.stdout}')
+            logger.debug(f'ClustalO stderr is: {result.stderr}')
+
+        except subprocess.CalledProcessError as exc:
+            logger.error(f'ClustalO FAILED. Output is: {exc}')
+            logger.error(f'ClustalO stdout is: {exc.stdout}')
+            logger.error(f'ClustalO stderr is: {exc.stderr}')
+
+            raise ValueError('There was an issue running ClustalO. Check input files!')
 
         with lock:
             counter.value += 1
@@ -503,8 +583,16 @@ def fasttree_multiprocessing(alignments_folder,
                                       bootstraps=bootstraps,
                                       logger=logger)
                           for alignment in alignments]
-        for future in future_results:
-            future.add_done_callback(utils.done_callback)
+
+        for future in as_completed(future_results):
+
+            try:
+                check = future.result()
+
+            except Exception as error:
+                print(f'Error raised: {error}')
+                tb = traceback.format_exc()
+                print(f'traceback is:\n{tb}')
 
         wait(future_results, return_when="ALL_COMPLETED")
 
@@ -619,8 +707,16 @@ def iqtree_multiprocessing(alignments_folder,
                                       bootstraps=bootstraps,
                                       logger=logger)
                           for alignment in alignments]
-        for future in future_results:
-            future.add_done_callback(utils.done_callback)
+
+        for future in as_completed(future_results):
+
+            try:
+                check = future.result()
+
+            except Exception as error:
+                print(f'Error raised: {error}')
+                tb = traceback.format_exc()
+                print(f'traceback is:\n{tb}')
 
         wait(future_results, return_when="ALL_COMPLETED")
 
@@ -730,9 +826,6 @@ def main(args,
                              qc_alignments_directory: qc_alignments_suffix}
     file_list = []
 
-    # if args.external_outgroups_file:
-    #     file_list.append(args.external_outgroups_file)
-
     utils.check_inputs(directory_suffix_dict,
                        file_list,
                        logger=logger)
@@ -765,13 +858,7 @@ def main(args,
         sys.exit()
 
     # Add outgroup sequences, both internal (if removed by the tree QC steps) and external (if a fasta file of
-    # external outgroup sequences is provided).
-    # outgroups_added_folder = add_outgroup_seqs(args.qc_alignment_directory,
-    #                                            selected_alignments_directory,
-    #                                            args.internal_outgroups,
-    #                                            args.external_outgroups_file,
-    #                                            list_of_external_outgroups=args.external_outgroups,
-    #                                            logger=logger)
+    # external outgroup sequences is provided):
 
     outgroups_added_folder = add_outgroup_seqs(args.qc_alignment_directory,
                                                selected_alignments_directory,